Task2/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-multipart
python-dotenv
openpyxl
python-jose
passlib
pytz
SQLAlchemy
asyncpg
requests